+++ ClubDoorman.Test/TestKit/TestKit.MessageHandlerBuilder.cs
@@ -6,6 +6,7 @@ using ClubDoorman.Models;
 using ClubDoorman.Models.Notifications;
 using ClubDoorman.Models.Requests;
 using ClubDoorman.Test.TestInfrastructure;
+using ClubDoorman.Test.TestKit.Builders.MockBuilders;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
